+++ ios/Hexter/Bridge/Hexter-Bridging-Header.h
@@ -1,0 +1,1 @@
+#include "hexter_bridge.h"